Update VK platform code according to latest framework changes Add dependency to run build on Windows Add IDE's folders to gitignore
