+++ packages/platform-vk/src/context.ts
@@ -1,40 +1,65 @@
 import type { IMessageContextSendOptions, MessageContext } from 'vk-io';
-import { PlatformContext } from '@henta/core';
-import type HentaBot from '@henta/core';
+import type { ISendMessageOptions } from '@henta/core';
+import { normalizeUploads, PlatformContext } from '@henta/core';
 import getKeyboardButton from './util/keyboard.js';
 import VkAttachment from './attachment.js';
-import type PlatformVk from './index.js';
-import type { ISendMessageOptions } from '@henta/core';
-import { normalizeUploads } from '@henta/core';
 import { uploadFile } from './util/files.js';
 
 export default class PlatformVkContext extends PlatformContext {
-  source = 'vk';
-  declare raw: MessageContext;
-  declare platform: PlatformVk;
+  public readonly source = 'vk';
+  public declare raw: MessageContext;
 
-  constructor(raw: MessageContext, bot: HentaBot, platform: any) {
-    super(raw, bot, platform);
+  public constructor(raw: MessageContext, platform: any) {
+    super(raw, null, platform);
     this.text = this.raw.text;
+    this.payload = this.raw.messagePayload;
   }
 
-  get originalText() {
+  public get originalText() {
     return this.raw.text;
   }
 
-  get peerId(): string {
-    return this.raw.peerId.toString();
-  }
-
-  get senderId() {
+  public get senderId() {
     return this.raw.senderId.toString();
   }
 
-  serialize() {
+  public get isChat() {
+    return this.raw.isChat;
+  }
+
+  public get peerId(): string {
+    return this.raw.peerId.toString();
+  }
+
+  public get attachments() {
+    return this.raw.attachments.map(
+      (attachment) =>
+        new VkAttachment(attachment.type, attachment.toJSON(), this.platform),
+    );
+  }
+
+  public get nestedAttachments() {
+    const response = [];
+
+    if (this.raw.hasReplyMessage) {
+      response.push(...this.raw.replyMessage.attachments);
+    }
+
+    if (this.raw.hasForwards) {
+      this.raw.forwards.forEach((v) => response.push(...v.attachments));
+    }
+
+    return response.map(
+      (attachment) =>
+        new VkAttachment(attachment.type, attachment.toJSON(), this.platform),
+    );
+  }
+
+  public serialize() {
     return this.raw['payload'];
   }
 
-  async send(message: ISendMessageOptions, isAnswer = false) {
+  public async send(message: ISendMessageOptions, isAnswer = false) {
     let attachment = [];
     if (message.files?.length) {
       const files = await normalizeUploads(message.files);
@@ -83,37 +108,5 @@ export default class PlatformVkContext extends PlatformContext {
     }
 
     return this.raw.send(messageBody);
-  }
-
-  get payload() {
-    return this.raw.messagePayload;
-  }
-
-  get isChat() {
-    return this.raw.isChat;
-  }
-
-  get attachments() {
-    return this.raw.attachments.map(
-      (attachment) =>
-        new VkAttachment(attachment.type, attachment.toJSON(), this.platform),
-    );
-  }
-
-  get nestedAttachments() {
-    const response = [];
-
-    if (this.raw.hasReplyMessage) {
-      response.push(...this.raw.replyMessage.attachments);
-    }
-
-    if (this.raw.hasForwards) {
-      this.raw.forwards.forEach((v) => response.push(...v.attachments));
-    }
-
-    return response.map(
-      (attachment) =>
-        new VkAttachment(attachment.type, attachment.toJSON(), this.platform),
-    );
   }
 }
