+++ packages/platform-vk/src/platform/updates/updates.behaviour.ts
@@ -1,0 +1,8 @@
+import { UpdatesBehaviour } from '@henta/core';
+import type VkPlatform from '..';
+
+export default abstract class VkUpdatesBehaviour extends UpdatesBehaviour {
+  public constructor(protected readonly platform: VkPlatform) {
+    super();
+  }
+}
