+++ packages/platform-vk/src/types/options.interface.ts
@@ -1,0 +1,9 @@
+export default interface IVKPlatformOptions {
+  token: string;
+  webhookConfirmation: string;
+  webhookSecret: string;
+  groupId: number;
+
+  /** @default longpoll */
+  updatesMode?: 'longpoll';
+}
